--- a/src/main/java/org/apache/sysds/lops/Binary.java
+++ b/src/main/java/org/apache/sysds/lops/Binary.java
@@ -38,6 +38,8 @@ public class Binary extends Lop
 {
 	private OpOp2 operation;
 	private final int _numThreads;
+
+	private final boolean inplace;
 	
 	/**
 	 * Constructor to perform a binary operation.
@@ -55,9 +57,14 @@ public class Binary extends Lop
 	}
 	
 	public Binary(Lop input1, Lop input2, OpOp2 op, DataType dt, ValueType vt, ExecType et, int k) {
+		this(input1, input2, op, dt, vt, et, k, false);
+	}
+
+	public Binary(Lop input1, Lop input2, OpOp2 op, DataType dt, ValueType vt, ExecType et, int k, boolean inplace) {
 		super(Lop.Type.Binary, dt, vt);
 		init(input1, input2, op, dt, vt, et);
 		_numThreads = k;
+		this.inplace = inplace; 
 	}
 	
 	private void init(Lop input1, Lop input2, OpOp2 op, DataType dt, ValueType vt, ExecType et)  {
@@ -106,6 +113,11 @@ public class Binary extends Lop
 			ret = InstructionUtils.concatOperands(ret, String.valueOf(_numThreads));
 		else if( getExecType() == ExecType.FED )
 			ret = InstructionUtils.concatOperands(ret, String.valueOf(_numThreads), _fedOutput.name());
+
+		if (getExecType() == ExecType.CP && inplace){
+			ret = InstructionUtils.concatOperands(ret, "InPlace");
+		}
+
 
 		return ret;
 	}