--- a/src/main/java/org/apache/sysds/runtime/compress/colgroup/AColGroupValue.java
+++ b/src/main/java/org/apache/sysds/runtime/compress/colgroup/AColGroupValue.java
@@ -204,6 +204,11 @@ public abstract class AColGroupValue extends ADictBasedColGroup {
 	}
 
 	@Override
+	public void clear(){
+		counts = null;
+	}
+
+	@Override
 	public String toString() {
 		StringBuilder sb = new StringBuilder();
 		sb.append(super.toString());