… src/main/java/org/apache/sysds/runtime/compress/colgroup/offset/AOffset.java
… src/main/java/org/apache/sysds/runtime/compress/colgroup/offset/OffsetSingle.java
… src/main/java/org/apache/sysds/runtime/compress/colgroup/offset/OffsetTwo.java
… src/main/java/org/apache/sysds/runtime/compress/lib/CLALibBinaryCellOp.java
… src/main/java/org/apache/sysds/runtime/compress/lib/CLALibDecompress.java
… src/main/java/org/apache/sysds/runtime/compress/lib/CLALibRightMultBy.java
… src/main/java/org/apache/sysds/runtime/compress/readers/ReaderColumnSelectionSparseTransposed.java
… src/main/java/org/apache/sysds/runtime/compress/utils/CompressRDDClean.java
… src/main/java/org/apache/sysds/runtime/controlprogram/context/SparkExecutionContext.java
… src/main/java/org/apache/sysds/runtime/instructions/Instruction.java
… src/main/java/org/apache/sysds/runtime/instructions/cp/BinaryCPInstruction.java
… src/main/java/org/apache/sysds/runtime/instructions/cp/BinaryMatrixMatrixCPInstruction.java
… src/main/java/org/apache/sysds/runtime/instructions/spark/functions/ExtractBlockForBinaryReblock.java
… src/main/java/org/apache/sysds/runtime/instructions/spark/utils/SparkUtils.java
… src/main/java/org/apache/sysds/runtime/matrix/data/LibMatrixBincell.java
… src/main/java/org/apache/sysds/runtime/matrix/data/LibMatrixReorg.java
… src/test/java/org/apache/sysds/test/component/compress/offset/LargeOffsetTest.java
… src/test/java/org/apache/sysds/test/component/compress/readers/ReadersTest.java